+++ src/datasource/CampaignDatasource.ts
@@ -1,21 +1,18 @@
 import { BASE_URL, httpClient } from "../config/develop-config";
 import { CreateCampaignEntiry } from "../entities/CampaignPointEntites";
 import { MissionDetailEntity } from "../entities/MissionEntities";
-import {
-  DetailRedeemDronerEntity,
-  RedeemDronerListEntity,
-} from "../entities/RedeemEntities";
+import { RedeemDronerListEntity } from "../entities/RedeemEntities";
 
 export class CampaignDatasource {
   static getCampaignList(
-    campaignType: string,
-    take: number,
-    page: number,
+    app?: string,
+    campaignType?: string,
+    take?: number,
+    page?: number,
     startDate?: string,
     endStart?: string,
     status?: string,
-    search?: string,
-    app?: string
+    search?: string
   ) {
     const params = {
       campaignType: campaignType,
@@ -148,7 +145,6 @@ export class CampaignDatasource {
   ) {
     const params = {
       application: app,
-
       take: take,
       page: page,
       startDate: startDate,
@@ -158,6 +154,69 @@ export class CampaignDatasource {
     };
     return httpClient
       .get(BASE_URL + `/promotion/campaign/find-all-campaign-quota`, { params })
+      .then((res) => {
+        return res.data;
+      })
+      .catch((err) => console.log(err));
+  }
+  static createCampaignQuota(
+    data: any,
+    coverFile: any,
+    floatingFile: any,
+    rewardFile: any,
+    rewardRoundFile: any
+  ) {
+    const formData = new FormData();
+    formData.append("campaignName", data.campaignName);
+    formData.append("campaignType", data.campaignType);
+    formData.append("startDate", data.startDate);
+    formData.append("endDate", data.endDate);
+    formData.append("condition", data.condition);
+    formData.append("createBy", data.createBy);
+    formData.append("updateBy", data.updateBy);
+    formData.append("status", data.status);
+    formData.append("application", data.application);
+    formData.append("rulesCampaign", data.rulesCampaign);
+    formData.append("description", data.description);
+    formData.append("floatingFile", floatingFile.file);
+    formData.append("rewardFile", rewardFile.file);
+    formData.append("rewardRoundFile", rewardRoundFile.file);
+    formData.append("coverFile", coverFile.file);
+    return httpClient
+      .post(BASE_URL + `/promotion/campaign/create-campaign-quota`, formData)
+      .then((res) => {
+        return res.data;
+      })
+      .catch((err) => console.log(err));
+  }
+  static updateCampaignQuota(
+    id: string,
+    data: any,
+    coverFile: any,
+    floatingFile: any,
+    rewardFile: any,
+    rewardRoundFile: any
+  ) {
+    const formData = new FormData();
+    formData.append("campaignName", data.campaignName);
+    formData.append("campaignType", data.campaignType);
+    formData.append("startDate", data.startDate);
+    formData.append("endDate", data.endDate);
+    formData.append("condition", data.condition);
+    formData.append("updateBy", data.updateBy);
+    formData.append("status", data.status);
+    formData.append("application", data.application);
+    formData.append("rulesCampaign", data.rulesCampaign);
+    formData.append("description", data.description);
+    formData.append("floatingFile", floatingFile.file);
+    formData.append("rewardFile", rewardFile.file);
+    formData.append("rewardRoundFile", rewardRoundFile.file);
+    formData.append("coverFile", coverFile.file);
+    return httpClient
+      .patch(
+        BASE_URL + `/promotion/campaign/update-campaign-quota/${id}`,
+        formData
+      )
       .then((res) => {
         return res.data;
       })