+++ src/definitions/Status.ts
@@ -150,7 +150,7 @@ export const STATUS_COUPON: any = {
   INACTIVE: "ปิดการใช้งาน",
 };
 export const REWARD_STATUS = [
-  { value: "", name: "ทั้งหมด" },
+  //{ value: "", name: "ทั้งหมด" },
   { value: "ACTIVE", name: "ใช้งาน" },
   { value: "DRAFTING", name: "รอเปิดใช้งาน" },
   { value: "INACTIVE", name: "ปิดการใช้งาน" },