+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## [Version - 0]
+## improve [v0.34.7] 2023-06-21
+- added soring in point menu
+- connected api create & edit & delete qtota
 ## improve [v0.34.6] 2023-06-19
 - enhance news menu
 - added category news & challenge